Remove partial at-most-once semantic
Currently, the client (browser) never retries any request making any
implementation of at-most-once unnecessary. In addition, in our use case
(powerful enough hardware for a proper browser which is connected via
LAN to the server) at-most-once is not really required as we can expect
to get no temporary failures which can be retried later.

Remove the token columns and all functionality using it to simplify the
code.