Remove SerializationError and all related code
The code used SerializationError and filterOperationalError() to catch
psycopg2.extensions.TransactionRollbackError errors which can occur with
ISOLATION_LEVEL_SERIALIZABLE.

However, instead of retrying the transaction (or the function) the code
just used the SerializationError exception to inform the user, that
"something is wrong". This is not useful and is also served by the
generic error which is displayed when Django fails with an exception.

Do exactly that and remove all handling of TransactionRollbackError. If
this errors occurs in production reintroduce the checks and implement
proper retries.