admin: prevent changes/deletions for Charge, Purchase, Transfer, Login
Admins should have no need to modify or delete existing objects of these
models. It also prevents operations which cause unsound state in the
database (e.g. deleting a charge but not updating the balance in the
userdata of the charged user).

Accidental operations can be reverted by simply adding the inverse
operation as new object.