backend: simply return recent users in getFrequentUsersList()
The old return value was complicated and I think it doesn't add much.
When there are many users the list will be too long anyway, so users
will either have to search a long list or type in their id.

Simplify that by just sorting the list after recent login.