backend: fix possible database race conditions
I'm not sure if all of these can actual trigger in reality because
Django obfuscates the time when data is loaded from the database. For
example in TransferLogic the race in money of receiver/sender was only
possible if the lines "receiver = transfer.receiver" and "sender =
transfer.sender" cause a database query. However, play it safe and move
all possible loads inside a transaction.

See also b08be91 ("admin: fix database race condition in ChargeAdmin and
TransferAdmin", 2019-12-05).